+++ src/lermitage/intellij/extra/icons/ExtraIconProvider.java
@@ -41,7 +41,7 @@ public class ExtraIconProvider extends IconProvider {
                 m("/icons/codeship.png").eq("codeship-steps").mayEnd(cfg),
                 m("/icons/contact.png").eq("contact", "contacts").mayEnd(txt),
                 m("/icons/contributing.png").eq("contribute", "contributing", "contribution").mayEnd(txt),
-                m("/icons/docker.png").eq("dockerfile", "docker-compose").mayEnd(yml),
+                m("/icons/docker.png").eq("dockerfile", "docker-compose", ".dockerignore").mayEnd(yml),
                 m("/icons/editorconfig.png").eq(".editorconfig"),
                 m("/icons/email.png").eq(".mailmap"),
                 m("/icons/git.png").eq(".gitattributes", ".gitignore", ".gitmodules"),