+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * detect more Docker and Docker Compose files.
 * support [Zappr](https://zappr.opensource.zalan.do) `.zappr.yaml` files.
 * support [Codecov](https://docs.codecov.io/docs/codecov-yaml) `codecov.yml`, `.codecov.yml` files. 
+* support `.dockerignore` files.
 
 ### 0.12 (2018/10/10)
 * reworked Gitlab files (reduced margin by 1px).